+++ src/math/Vec2.ts
@@ -199,6 +199,17 @@ export class Vec2 {
 	}
 
 	/**
+	 * アダマール積を求める。
+	 *
+	 * @param v ベクトル。
+	 */
+	mul(v: Vec2Like): this {
+		this.x = this.x * v.x;
+		this.y = this.y * v.y;
+		return this;
+	}
+
+	/**
 	 * 内積を求める。
 	 *
 	 * @param v ベクトル。
@@ -286,6 +297,42 @@ export class Vec2 {
 	negate(): this {
 		this.x *= -1;
 		this.y *= -1;
+		return this;
+	}
+
+	/**
+	 * 四捨五入する。
+	 */
+	round(): this {
+		this.x = Math.round(this.x);
+		this.y = Math.round(this.y);
+		return this;
+	}
+
+	/**
+	 * 小数点以下を切り捨てる。
+	 */
+	floor(): this {
+		this.x = Math.floor(this.x);
+		this.y = Math.floor(this.y);
+		return this;
+	}
+
+	/**
+	 * 小数点以下を切り上げる。
+	 */
+	ceil(): this {
+		this.x = Math.ceil(this.x);
+		this.y = Math.ceil(this.y);
+		return this;
+	}
+
+	/**
+	 * 各要素を符号を表す +/- 1, +/- 0 にする。
+	 */
+	sign(): this {
+		this.x = Math.sign(this.x);
+		this.y = Math.sign(this.y);
 		return this;
 	}