+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## Unreleased changes
+
+- `Vec2`, `Vec3` クラスに `mul()`, `round()`, `floor()`, `ceil()`, `sign()` を追加。
+
 ## 1.0.1
 
 - ドキュメンテーションコメント中、角度の単位をradian出なくdegreeと誤って説明している箇所の修正。